Refactor app: fix hook, context imports
