+++ src/components/TodosList.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import { setTodoFilter, setTodos } from '../slices/todoReducer';
 import selectTodoItems, { selectTodoFilter } from '../slices/todoSelector';
-import useLocalStorage from '../hooks/index.js';
+import { useLocalStorage } from '../hooks/index.js';
 
 import Todo from './Todo';
